Change ALLOWED\_HOSTS to allow any host in debug mode
This commit changes the ALLOWED\_HOSTS setting in the Django project's settings.py file from an empty list to a list containing a wildcard character (\*). This allows any host to access the site, but only when the DEBUG setting is set to True. This is useful for local development and testing, but should never be used in a production environment.