Added static URL pattern for serving static files in debug mode
In the given code diff, a new block of code has been added to the bottom of the `urlpatterns` list in the `urls.py` file located in the `notes` directory. This new block checks if the Django project is in debug mode using `settings.DEBUG`. If it is, it adds a new URL pattern that maps the static URL (`settings.STATIC_URL`) to the static root directory (`settings.STATIC_ROOT`). This allows static files to be served during development.

The `static` function is imported from `django.conf.urls.static` and the `settings` module is imported from `django.conf` at the beginning of the file.