Change static files storage to use Whitenoise compression
In this diff, the line that sets `STATICFILES_STORAGE` to use Whitenoise compression has been uncommented in `settings.py`. Previously, this line was commented out and there was a section of code defining `STORAGES` which included a similar configuration. However, this section of code has now been commented out.

Overall, this diff enables compression of static files using Whitenoise.